backend/requirements.txt
# Core data processing
numpy>=1.26.1
pandas>=1.3.3
scikit-learn>=1.0.0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=1.4.0
pymongo>=3.12.0
python-dotenv>=0.19.0

# Data validation
pydantic>=2.4.2
pydantic-settings==2.0.3

# Utilities
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0

# Development
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0

# Image processing
pillow==10.1.0

# Audio processing
librosa==0.10.1

# Text processing
nltk==3.8.1
soundfile==0.12.1

# 3D processing
trimesh==4.0.5
